Add auto play function
The playback can be auto started/stoped when the tracking is toggled. Currently the implementation relies on the qt_viewer interface for the playing. This might have to be changed in the futur since public access to  will be removed in napari v0.5.0. I will open an issue on github to discuss alternatives.