Add layers that existed before the plugin was opened to drop-down menus
Images and point layers that were open pefore the plugin was launched were previously not added to the drop-down menus. This required some refactoring of the call backs from insertion and deletion of layers such that the functions can be used to add/remove an item from the menu direction without passing an event.