Add notification for missing points layer and initialize the playback parameters correctly
When the users started tracking without a points layer an error for missing '' key in dictionary was raised but the playback started regardless. This could have left the user confused because it looks like the tracking is working but the points could not be saved and the error message is not clear as to what the problem is. Now an error is raised telling the user to selct/creat a points layer for tracking.

This commit also initialized the playback parameters with the defaults in the playback parameters QGroupBox. Previously napari just used the parameters form the last time napari was open. This can create a discrepancy between the parameters used and the ones shown in the QGroupBox.