Fix lage when tracking by scrolling
Previously the the plugin updated the points layer of the currently shown frame when scrolling through the images. This is a problem, because you want to position the cursor on a given frame and then save it by scrolling to the next frame, i.e. the cursor position when changing frame should be saved for the frame prior to the transition and not post. To ensure that this is handled correctly when scrolling backwards and forwards, I introduced a new property to the widget class that keeps track of the previous frame ().